+++ packages/analytics-js-integrations/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.24.2](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-integrations@3.24.1...@rudderstack/analytics-js-integrations@3.24.2) (2026-08-13)
+
+
+### Bug Fixes
+
+* **@rudderstack/analytics-js-integrations:** relax Google Ads identify validation ([#3152](https://github.com/rudderlabs/rudder-sdk-js/issues/3152)) ([d4adcfd](https://github.com/rudderlabs/rudder-sdk-js/commit/d4adcfdea534b37bbe45792336cd540b700f0b7a))
+
 ## [3.24.1](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-integrations@3.24.0...@rudderstack/analytics-js-integrations@3.24.1) (2026-08-05)