+++ packages/analytics-v1.1/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [2.52.18](https://github.com/rudderlabs/rudder-sdk-js/compare/rudder-sdk-js@2.52.17...rudder-sdk-js@2.52.18) (2026-08-13)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-integrations` updated to version `3.24.2`
+
 ## [2.52.17](https://github.com/rudderlabs/rudder-sdk-js/compare/rudder-sdk-js@2.52.16...rudder-sdk-js@2.52.17) (2026-08-05)
 
 ### Dependency Updates